mapic: Rename stats -> metrics on our side
Mist calls them stats, but I think that in our
context they are more like metrics about the
current state, streams and pushes (and later
ingest, etc). So I kept it stats whenever it
was a mist thing, but renamed the event and
any intermediate concepts to metrics as well.