mist-api-connector: Export mist stats to RabbitMQ (#104)
* apis/mist: Create API to grab stream stats

Included pushes stats as well just because it
was just there.

* mapic: Create stats collector pulling from Mist

Really simple logic just to show what I have in my mind.
LMK any early feedback! I could move it to the mistapiconnector_app
as well, but I just thought it was a logic so separate from the
main flow that it was worth a separate component.

* mapic: Use active_streams command for stats as well

Also a few improvements to JSON unmarshalling of
response:
 - Make unmarshalling all or nothing
 - Allow push stats to be nil to know when it's not present
 - Remove unknown field from struct, keep it on parsing logic

* mapic: Revert back to stats_streams

They actually work differently when we enable HLS sessions.

* mapic: Use lower-level AMQPProducer

Will allow flexibly publishing to different exchanges.

* mapic: Send stream stats event to rabbitmq

* mapic: Final touches and plug collector to *mac

* WIP: A couple reminder todos

* mapic: Implement the missing corner case TODOs

* mapic: Make sure we don't panic crash on stats collector

Create a separate function so we can do an easy defer recover
in there and keep looping. that log should be picked by any
panic alarm we eventually create.

* mapic: Rename stats -> metrics on our side

Mist calls them stats, but I think that in our
context they are more like metrics about the
current state, streams and pushes (and later
ingest, etc). So I kept it stats whenever it
was a mist thing, but renamed the event and
any intermediate concepts to metrics as well.

* Tune metrics event topic

* go.mod: livepeer-data version 0.3.0-beta.2

* mapic: Remove viewer count stuff from metrics event

* apis/mist: Fix info log for stats API