.github/ci: Build images on every branch
Use the always-latest-on-branch input to the gh action
so we only publish `latest` tags on the specified branches.

While `latest` will only be pushed on those branches, we
can still test the generated images from any branch since
I removed the action filter to only run on master/mapic branhc.