recordtester: Add ingest arg to skip LP /api/ingest
That API now returns the global ingest and we don't want that.
We should instead stream into the region being tested and so
the best way for that should be receiving an explicit argument
instead, since we don't want to expose the regional URLs in any
way to our users.