record-tester: Fix stream health always firing with other errs
We were passing an always not nil error to sendPagerDutyEvent.

We could make the error a pointer, but that wouldn't fix anything
since when we convert StreamHealthError to error it would still
result in a non-nil error: https://go.dev/play/p/1JDB0I910-4

So just drop the variable with an intermediate type, and set it
based on the result of errors.As instead of by errors.As.